# docs/swagger.yaml
basePath: /
definitions:
  github_com_actanonvebra_honeyshop_internal_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: ibrahimserhatbulut@gmail.com
    name: actanonvebra
  description: This is the API documentation for the Honeyshop honeypot project.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Honeyshop API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return token
      parameters:
      - description: User's login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/github_com_actanonvebra_honeyshop_internal_models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_actanonvebra_honeyshop_internal_models.User'
        "400":
          description: Invalid credentials
          schema:
            type: string
        "429":
          description: Too many requests
          schema:
            type: string
      summary: Login User
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User's Username
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: User's Password
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: User's Email
        in: body
        name: email
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_actanonvebra_honeyshop_internal_models.User'
        "400":
          description: Missing or invalid input
          schema:
            type: string
      summary: Register User
      tags:
      - auth
swagger: "2.0"